% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-sdy-extensions.R
\name{coefficient_of_determination}
\alias{coefficient_of_determination}
\title{Coefficient of determination}
\usage{
coefficient_of_determination(validity)
}
\arguments{
\item{validity}{Predictor-criterion validity coefficient.}
}
\value{
Numeric vector with \code{validity^2}.
}
\description{
Computes the squared validity coefficient.
}
\examples{
# Literature: Taylor and Russell (1939); Holling (1998).
coefficient_of_determination(.30)
}
\references{
Taylor, H. C., & Russell, J. T. (1939). The relationship of validity
coefficients to the practical effectiveness of tests in selection. Journal
of Applied Psychology, 23, 565-578.

Holling, H. (1998). Utility analysis of personnel selection: An overview and
empirical study based on objective performance measures. Methods of
Psychological Research Online, 3(1), 5-24.
}
