% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{d_to_cor}
\alias{d_to_cor}
\title{Convert Cohen's d to a correlation}
\usage{
d_to_cor(d)
}
\arguments{
\item{d}{Cohen's d.}
}
\value{
Correlation coefficient.
}
\description{
Uses \eqn{r = d / \sqrt{d^2 + 4}}.
}
\examples{
# Literature: Cohen (1988); Schmidt, Hunter, and Pearlman (1982).
d_to_cor(.50)
}
\references{
Schmidt, F. L., Hunter, J. E., & Pearlman, K. (1982). Assessing the economic
impact of personnel programs on workforce productivity. Personnel
Psychology, 35, 333-347.

Cohen, J. (1988). Statistical power analysis for the behavioral sciences
(2nd ed.). Erlbaum.
}
