% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-sdy-extensions.R
\name{forecasting_efficiency}
\alias{forecasting_efficiency}
\title{Forecasting efficiency}
\usage{
forecasting_efficiency(validity)
}
\arguments{
\item{validity}{Predictor-criterion validity coefficient.}
}
\value{
Numeric vector with forecasting efficiency values.
}
\description{
Computes the proportional reduction in the standard error of prediction:
\code{1 - sqrt(1 - validity^2)}.
}
\examples{
# Literature: Holling (1998).
forecasting_efficiency(.30)
}
\references{
Holling, H. (1998). Utility analysis of personnel selection: An overview and
empirical study based on objective performance measures. Methods of
Psychological Research Online, 3(1), 5-24.
}
