% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse-helpers.R
\name{fuse_reliability}
\alias{fuse_reliability}
\title{Reliability of a weighted composite (Mosier, 1943; Lord & Novick, 1968)}
\usage{
fuse_reliability(weights, item_cor, item_reliabilities = NULL)
}
\arguments{
\item{weights}{Numeric vector of composite weights.}

\item{item_cor}{Symmetric correlation (or covariance) matrix among items.}

\item{item_reliabilities}{Numeric vector of item reliabilities (length
equal to \code{weights}). If \code{NULL}, the composite reliability is computed
under the assumption that diagonal entries of \code{item_cor} are item
reliabilities (e.g., when an empirical reliability matrix is supplied).}
}
\value{
The reliability of the weighted composite.
}
\description{
Reliability of a weighted composite (Mosier, 1943; Lord & Novick, 1968)
}
\examples{
R <- matrix(c(1, .3, .3, 1), 2, 2)
fuse_reliability(c(.5, .5), R, item_reliabilities = c(.80, .85))
}
\references{
Lord, F. M., & Novick, M. R. (1968). \emph{Statistical theories of mental test
scores}. Addison-Wesley.

Mosier, C. I. (1943). On the reliability of a weighted composite.
\emph{Psychometrika}, 8, 161-168.
}
