% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse-helpers.R
\name{fuse_validity}
\alias{fuse_validity}
\title{Correlation of a weighted composite with an external variable}
\usage{
fuse_validity(weights, item_cor, item_validities)
}
\arguments{
\item{weights}{Composite weights.}

\item{item_cor}{Predictor (item) correlation matrix.}

\item{item_validities}{Item-level correlations with the external variable.}
}
\value{
Scalar correlation.
}
\description{
Implements the standard formula
\eqn{r_{C,Y} = (w' \rho_{XY}) / \sqrt{w' R_{XX} w}}
for the correlation between a weighted composite of items and an external
criterion \code{Y}, where the items have correlations \code{R_XX} and individual
validities \verb{\\rho_\{XY\}} (Lord & Novick, 1968, Ch. 4).
}
\examples{
R <- matrix(c(1, .3, .3, 1), 2, 2)
fuse_validity(c(.5, .5), R, item_validities = c(.30, .25))
}
