% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{inflation_adjusted_rate}
\alias{inflation_adjusted_rate}
\title{Combine nominal discount and inflation rates}
\usage{
inflation_adjusted_rate(discount_rate, inflation_rate)
}
\arguments{
\item{discount_rate}{Real discount rate.}

\item{inflation_rate}{Inflation rate.}
}
\value{
Inflation-adjusted discount rate.
}
\description{
Computes \code{i_a = i + f + i*f}.
}
\examples{
# Literature: Tziner et al. (1994); Holling (1998).
inflation_adjusted_rate(.08, .025)
}
\references{
Tziner, A., Meir, E. I., Dahan, M., & Birati, A. (1994). An investigation of
the predictive validity and economic utility of the assessment center for
the high- management level. Canadian Journal of Behavioural Science, 26,
228-245.

Holling, H. (1998). Utility analysis of personnel selection: An overview and
empirical study based on objective performance measures. Methods of
Psychological Research Online, 3(1), 5-24.
}
