% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiattribute-fairness.R
\name{pareto_frontier}
\alias{pareto_frontier}
\title{Pareto frontier indicator}
\usage{
pareto_frontier(objectives, maximize = TRUE)
}
\arguments{
\item{objectives}{Numeric matrix/data frame. Alternatives are rows, objectives columns.}

\item{maximize}{Logical vector indicating whether each objective is to be maximized.
Scalar values are recycled.}
}
\value{
Logical vector indicating Pareto-efficient rows.
}
\description{
Identifies non-dominated alternatives for objectives to be maximized or minimized.
}
\examples{
# Literature: De Corte, Lievens, and Sackett (2007); De Corte, Sackett, and Lievens (2011).
pareto_frontier(data.frame(validity = c(.30, .35, .32), diversity = c(.80, .70, .85)))
}
\references{
De Corte, W., Lievens, F., & Sackett, P. R. (2007). Combining predictors to
achieve optimal trade-offs between selection quality and adverse impact.
Journal of Applied Psychology, 92, 1380-1393. De Corte, W., Sackett, P. R.,
& Lievens, F. (2011). Designing Pareto-optimal selection systems:
Formalizing the decisions required for selection system development. Journal
of Applied Psychology, 96, 907-926.
}
