% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdy.R
\name{sdy_percentile}
\alias{sdy_percentile}
\title{Estimate SDy from percentile judgments}
\usage{
sdy_percentile(p15, p85)
}
\arguments{
\item{p15}{Estimated monetary value of performance at the 15th percentile.}

\item{p85}{Estimated monetary value of performance at the 85th percentile.}
}
\value{
Estimated standard deviation of job performance in monetary units.
}
\description{
Implements the percentile approximation \code{SDy = (P85 - P15) / 2}.
}
\examples{
# Literature: Schmidt et al. (1979); Bobko, Karren, and Parkington (1983).
sdy_percentile(p15 = 60000, p85 = 140000)
}
\references{
Bobko, P., Karren, R., & Parkington, J. J. (1983). Estimation of standard
deviations in utility analyses: An empirical test. Journal of Applied
Psychology, 68, 170-176.

Schmidt, F. L., Hunter, J. E., McKenzie, R. C., & Muldrow, T. W. (1979).
Impact of valid selection procedures on work-force productivity. Journal of
Applied Psychology, 64, 609-626.
}
