% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdy.R
\name{sdy_proportional}
\alias{sdy_proportional}
\title{Estimate SDy with proportional rules}
\usage{
sdy_proportional(mean_pay, multiplier = 0.4)
}
\arguments{
\item{mean_pay}{Mean pay or mean output value.}

\item{multiplier}{Proportional SDy multiplier. Defaults to \code{.40}.}
}
\value{
Estimated SDy.
}
\description{
Computes a salary- or value-based SDy estimate using a multiplier such as .40
or .70.
}
\examples{
# Literature: Schmidt, Hunter, and Pearlman (1982); Hunter and Schmidt (1982).
sdy_proportional(mean_pay = 80000, multiplier = .40)
sdy_proportional(mean_pay = 80000, multiplier = .70)
}
\references{
Schmidt, F. L., Hunter, J. E., & Pearlman, K. (1982). Assessing the economic
impact of personnel programs on workforce productivity. Personnel
Psychology, 35, 333-347.

Hunter, J. E., & Schmidt, F. L. (1982). Fitting people to jobs: The impact
of personnel selection on national productivity. In M. D. Dunnette & E. A.
Fleishman (Eds.), Human performance and productivity (Vol. 1, pp. 233-284).
Erlbaum.
}
