% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdy.R
\name{sdy_rbn}
\alias{sdy_rbn}
\title{Estimate SDy with a coefficient-of-variation approach}
\usage{
sdy_rbn(mean_pay, coefficient_variation)
}
\arguments{
\item{mean_pay}{Mean pay or mean criterion value.}

\item{coefficient_variation}{Coefficient of variation for job performance value.}
}
\value{
Estimated SDy.
}
\description{
A compact implementation of the Raju-Burke-Normand logic: \code{SDy = CV * mean_pay}.
Use this function when the coefficient of variation is theoretically or
empirically justified for the job family.
}
\examples{
# Literature: Raju, Burke, and Normand (1990).
sdy_rbn(mean_pay = 80000, coefficient_variation = .35)
}
\references{
Raju, N. S., Burke, M. J., & Normand, J. (1990). A new approach for utility
analysis. Journal of Applied Psychology, 75, 3-12.
}
