% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced-decision-support.R
\name{sdy_superior_equivalents}
\alias{sdy_superior_equivalents}
\title{SDy from superior-equivalents judgments}
\usage{
sdy_superior_equivalents(superior_value, typical_value, z_difference = 1)
}
\arguments{
\item{superior_value}{Monetary value of the superior performer.}

\item{typical_value}{Monetary value of the typical performer.}

\item{z_difference}{Standardized distance between the two performers. Defaults
to \code{1}, but can be set to another value if the judgment anchors imply it.}
}
\value{
Estimated SDy.
}
\description{
Computes SDy from a judged monetary difference between a superior and a typical
performer, divided by the standardized distance assumed to separate them.
}
\examples{
# Literature: Eaton, Wing, and Mitchell (1985); Burke and Frederick (1984, 1986).
sdy_superior_equivalents(superior_value = 140000, typical_value = 100000)
}
\references{
Eaton, N. K., Wing, H., & Mitchell, K. J. (1985). Alternate methods of
estimating the dollar value of performance. Personnel Psychology, 38, 27-40.

Burke, M. J., & Frederick, J. T. (1984). Two modified procedures for
estimating standard deviations in utility analyses. Journal of Applied
Psychology, 69, 482-489.

Burke, M. J., & Frederick, J. T. (1986). A comparison of economic utility
estimates for alternative SDy estimation procedures. Journal of Applied
Psychology, 71, 334-339.
}
