% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{print.psu_utility}
\alias{print.psu_utility}
\title{Print personnel-selection utility objects}
\usage{
\method{print}{psu_utility}(x, ...)
}
\arguments{
\item{x}{An object returned by one of the package functions.}

\item{...}{Ignored.}
}
\description{
Print personnel-selection utility objects
}
\keyword{internal}
